+++ projects/nb-trans/src/lib/constants/nb-trans-lang.enum.ts
@@ -1,4 +1,4 @@
-export enum NbTransLangEnum {
+export enum NbTransLang {
   // 简体中文
   ZH_CN = 'zh-CN',
   // 繁体中文	
@@ -105,3 +105,7 @@ export enum NbTransLangEnum {
   // 越南语	
   VI_VN = 'vi-VN',
 }
+/**
+ * @deprecated use "NbTransLang" please
+ */
+export const NbTransLangEnum = NbTransLang;