+++ projects/nb-trans/src/lib/constants/nb-trans-sentence-item.enum.ts
@@ -1,5 +1,9 @@
-export enum NbTransSentenceItemEnum {
+export enum NbTransSentenceItem {
   STR,
   COMP,
   MULTI_COMP,
 };
+/**
+ * @deprecated use "NbTransSentenceItem" please
+ */
+export const NbTransSentenceItemEnum = NbTransSentenceItem;