+++ src/app/app.module.ts
@@ -1,6 +1,6 @@
 import { NgModule } from '@angular/core';
 import { BrowserModule } from '@angular/platform-browser';
-import { NbTransLangEnum, NbTransModule, NB_TRANS_DEFAULT_LANG, NB_TRANS_LOADER } from 'nb-trans';
+import { NbTransLang, NbTransModule, NB_TRANS_DEFAULT_LANG, NB_TRANS_LOADER } from 'nb-trans';
 
 import { AppRoutingModule } from './app-routing.module';
 import { AppComponent } from './app.component';
@@ -19,12 +19,12 @@ import { environment } from '../environments/environment';
   ],
   providers: [
     // {
-    //   provide: NG_TRANS_MAX_RETRY_TOKEN,
+    //   provide: NG_TRANS_MAX_RETRY,
     //   useValue: 0
     // },
     {
       provide: NB_TRANS_DEFAULT_LANG,
-      useValue: NbTransLangEnum.ZH_CN,
+      useValue: NbTransLang.ZH_CN,
     },
     // {
     //   provide: NG_TRANS_LOADER,
@@ -45,9 +45,9 @@ import { environment } from '../environments/environment';
 
         // dyn load and the content is a json file
         // [NgTransLangEnum.EN]: () => http.get('./assets/localization/en/translations.json').toPromise(),
-        [NbTransLangEnum.EN]: () => http.get(environment.domain + 'assets/localization/en/translations.json'),
+        [NbTransLang.EN]: () => http.get(environment.domain + 'assets/localization/en/translations.json'),
         // [NgTransLangEnum.ZH_CN]: () => http.get('./assets/localization/zh-CN/translations.json').toPromise(),
-        [NbTransLangEnum.ZH_CN]: () => http.get(environment.domain + 'assets/localization/zh-CN/translations.json'),
+        [NbTransLang.ZH_CN]: () => http.get(environment.domain + 'assets/localization/zh-CN/translations.json'),
       }),
       deps: [HttpClient]
     }